Refactor out work counters
